LayoutTests/imported/w3c/web-platform-tests/encoding/legacy-mb-tchinese/big5/big5-decode-errors-expected.txt


PASS lead not 0x00 and no more bytes: AB 
PASS lead not 0x00 and no more bytes: AB B5 AB 
PASS lead not 0x00 and no more bytes: AB AB AB 
FAIL lead byte outside 0x81-0xFE: FF assert_equals: expected "\ufffd" but got ""
FAIL lead byte outside 0x81-0xFE: AB AB FF assert_equals: expected "奏\ufffd" but got "奏"
PASS trail byte outside 0x41-0xFE: AB 31 
FAIL trail byte outside 0x41-0xFE: AB FF assert_equals: expected "\ufffd" but got "\ufffd"
FAIL pointer is null: C8 F2 assert_equals: expected "\ufffd" but got ""
